Update test_order_input.py
Updated testing code (correct me if im wrong). Since quite a bit of explaining done I have written most of it as comments in the code. One of the key changes I didnt comment is when you use pytest.raises(KeyError) you are telling pytest that you expect a KeyError to be raised by the code inside that block. If the code does not raise a KeyError, pytest will fail that test because the expectation was not met. But when using an assertion like assert result is False it checks the return value of the function. If order_book.add_order(missing_timestamp_order, account_manager) returns False instead of raising a KeyError, the test will pass because the assertion matches what the code actually does.